Update Clock.cs
Implemented Mentors suggestions

Good first try! My improvements are as follows:

Your while loops in Hours can be implemented much simpler with the % operator.
Similarly, Minutes can use modulo and division (rounding down) to achieve the same.
I also think that the properties should not have public setters, and that they should not do the computation, instead this should be done in methods such as Add.
Try to not implement adding via a DateTime object as it circumvents the point of not using dates.
Equals does not need the if statement, just return the expression.
Good luck!